[ALPS04808624] iommu: update trace log
[Detail]
1. improve ion DB log
2. correct iommu trace type of 34bit iova
3. allow user error handling of config buffer without mapping

MTK-Commit-Id: 8ab8a2e2cec50bdd0738972d5a51bef91274306f

CR-Id: ALPS04808624
Feature: [Module]ION/M4U

+++ drivers/iommu/mtk_iommu_ext.h
@@ -144,9 +144,9 @@ void mtk_iommu_trace_unmap(unsigned long orig_iova,
 			   size_t size,
 			   size_t unmapped);
 void mtk_iommu_trace_log(int event,
-			 unsigned int data1,
-			 unsigned int data2,
-			 unsigned int data3);
+			 unsigned long data1,
+			 unsigned long data2,
+			 unsigned long data3);
 
 void mtk_iommu_log_dump(void *seq_file);
 int m4u_user2kernel_port(int userport);
